Migrate BLRPRx from basemodel.py to blrprx.py.
