+++ src/pyBL/models/__init__.py
@@ -1,1 +1,2 @@
 from .basemodel import *
+from .blrprx import *
